pyproject.toml
[project]
name = "hands-on-whisper"
version = "0.1.0"
description = "Exploring faster-whisper and OpenAI whisper for speech recognition"
readme = "README.md"
license = "MIT"
authors = [{ name = "Damir Manapov" }]
requires-python = ">=3.11"
dependencies = [
  "faster-whisper>=1.1.0",
  "openai-whisper>=20240930",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "ruff>=0.8.6",
  "pip-audit>=2.7.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 2

[tool.ruff.lint]
select = [
  "E",      # pycodestyle errors
  "W",      # pycodestyle warnings
  "F",      # pyflakes
  "I",      # isort
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "UP",     # pyupgrade
  "ARG",    # flake8-unused-arguments
  "SIM",    # flake8-simplify
  "TCH",    # flake8-type-checking
  "PTH",    # flake8-use-pathlib
  "ERA",    # eradicate (commented out code)
  "PL",     # pylint
  "PERF",   # perflint
  "RUF",    # ruff-specific
  "ANN",    # flake8-annotations
]
ignore = [
  "PLR2004",  # magic value comparison (acceptable for CLI arg checks)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
